fixed foreignkey type tour_guides

+++ database/migrations/2019_09_23_062106_create_tour_guides_table.php
@@ -18,7 +18,7 @@ class CreateTourGuidesTable extends Migration
             $table->string('name');
             $table->string('email')->nullable(true);
             $table->string('phone')->nullable(true);
-            $table->unsignedInteger('tour_agent_id');
+            $table->unsignedBigInteger('tour_agent_id');
             $table->timestamps();
 
             $table->foreign('tour_agent_id')->references('id')->on('tour_agents')->onDelete('restrict');
